readtext

+++ file_io/0-read_textfile.c
@@ -1,7 +1,7 @@
 #include "main.h"
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-int lprint, lread, closed, file;
+size_t lprint, lread, closed, file;
 char *buffer;
 file = lprint = lread = 0;
 buffer = malloc (letters * sizeof(char));
@@ -28,7 +28,7 @@ closed = close(file);
 if (closed == -1)
 {
 return(0);
-}free(buffer);
+}
+free(buffer);
 return(lread);
-
 }
